# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "jobApplier API",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/sources/{key}": {
   "patch": {
    "summary": "Patch Source",
    "operationId": "patch_source_api_sources__key__patch",
    "parameters": [
     {
      "name": "key",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Key"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SourcePatch"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/boards": {
   "get": {
    "summary": "List Boards",
    "operationId": "list_boards_api_boards_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     }
    }
   },
   "post": {
    "summary": "Add Board",
    "operationId": "add_board_api_boards_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/BoardIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/boards/{bid}": {
   "patch": {
    "summary": "Patch Board",
    "operationId": "patch_board_api_boards__bid__patch",
    "parameters": [
     {
      "name": "bid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Bid"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/BoardPatch"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "summary": "Delete Board",
    "operationId": "delete_board_api_boards__bid__delete",
    "parameters": [
     {
      "name": "bid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Bid"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/applications": {
   "get": {
    "summary": "List Applications",
    "operationId": "list_applications_api_applications_get",
    "parameters": [
     {
      "name": "state",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "State"
      }
     },
     {
      "name": "min_score",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "number"
        },
        {
         "type": "null"
        }
       ],
       "title": "Min Score"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 200,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/applications/{app_id}": {
   "get": {
    "summary": "Application Detail",
    "operationId": "application_detail_api_applications__app_id__get",
    "parameters": [
     {
      "name": "app_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "App Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/applications/{app_id}/prepare": {
   "post": {
    "summary": "Prepare",
    "operationId": "prepare_api_applications__app_id__prepare_post",
    "parameters": [
     {
      "name": "app_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "App Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/applications/{app_id}/answers/{index}": {
   "post": {
    "summary": "Map Answer",
    "operationId": "map_answer_api_applications__app_id__answers__index__post",
    "parameters": [
     {
      "name": "app_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "App Id"
      }
     },
     {
      "name": "index",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Index"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AnswerMap"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/applications/{app_id}/approve": {
   "post": {
    "summary": "Approve",
    "operationId": "approve_api_applications__app_id__approve_post",
    "parameters": [
     {
      "name": "app_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "App Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ApproveIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/applications/{app_id}/submit": {
   "post": {
    "summary": "Submit",
    "description": "Runs the full deterministic gate. If anything fails, a handoff task is created instead.",
    "operationId": "submit_api_applications__app_id__submit_post",
    "parameters": [
     {
      "name": "app_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "App Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/applications/{app_id}/manual-submission": {
   "post": {
    "summary": "Manual Submission",
    "operationId": "manual_submission_api_applications__app_id__manual_submission_post",
    "parameters": [
     {
      "name": "app_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "App Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ManualSubmission"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/applications/{app_id}/reject": {
   "post": {
    "summary": "Reject",
    "operationId": "reject_api_applications__app_id__reject_post",
    "parameters": [
     {
      "name": "app_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "App Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/StateIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/applications/{app_id}/restore": {
   "post": {
    "summary": "Restore",
    "operationId": "restore_api_applications__app_id__restore_post",
    "parameters": [
     {
      "name": "app_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "App Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/StateIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/applications/{app_id}/confirm": {
   "post": {
    "summary": "Confirm",
    "operationId": "confirm_api_applications__app_id__confirm_post",
    "parameters": [
     {
      "name": "app_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "App Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/StateIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/applications/{app_id}/follow-up": {
   "post": {
    "summary": "Follow Up",
    "operationId": "follow_up_api_applications__app_id__follow_up_post",
    "parameters": [
     {
      "name": "app_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "App Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/StateIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/applications/{app_id}/resume.docx": {
   "get": {
    "summary": "Resume Docx",
    "operationId": "resume_docx_api_applications__app_id__resume_docx_get",
    "parameters": [
     {
      "name": "app_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "App Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/applications/{app_id}/cover-letter.txt": {
   "get": {
    "summary": "Cover Letter",
    "operationId": "cover_letter_api_applications__app_id__cover_letter_txt_get",
    "parameters": [
     {
      "name": "app_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "App Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/handoffs": {
   "get": {
    "summary": "List Handoffs",
    "operationId": "list_handoffs_api_handoffs_get",
    "parameters": [
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "OPEN",
       "title": "Status"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/handoffs/{hid}/dismiss": {
   "post": {
    "summary": "Dismiss Handoff",
    "operationId": "dismiss_handoff_api_handoffs__hid__dismiss_post",
    "parameters": [
     {
      "name": "hid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Hid"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "AnswerMap": {
    "properties": {
     "answer_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Answer Id"
     }
    },
    "type": "object",
    "required": [
     "answer_id"
    ],
    "title": "AnswerMap"
   },
   "ApproveIn": {
    "properties": {
     "answer_on_site": {
      "type": "boolean",
      "title": "Answer On Site",
      "default": false
     }
    },
    "type": "object",
    "title": "ApproveIn"
   },
   "BoardIn": {
    "properties": {
     "source_key": {
      "type": "string",
      "title": "Source Key"
     },
     "board_token": {
      "type": "string",
      "maxLength": 120,
      "minLength": 1,
      "title": "Board Token"
     },
     "employer_name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200
       },
       {
        "type": "null"
       }
      ],
      "title": "Employer Name"
     }
    },
    "type": "object",
    "required": [
     "source_key",
     "board_token"
    ],
    "title": "BoardIn"
   },
   "BoardPatch": {
    "properties": {
     "enabled": {
      "type": "boolean",
      "title": "Enabled"
     }
    },
    "type": "object",
    "required": [
     "enabled"
    ],
    "title": "BoardPatch"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ManualSubmission": {
    "properties": {
     "confirmation_id": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200
       },
       {
        "type": "null"
       }
      ],
      "title": "Confirmation Id"
     },
     "confirmation_url": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 600
       },
       {
        "type": "null"
       }
      ],
      "title": "Confirmation Url"
     }
    },
    "type": "object",
    "title": "ManualSubmission"
   },
...
   "SourcePatch": {
    "properties": {
     "enabled": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Enabled"
     },
     "auto_submit_opt_in": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Auto Submit Opt In"
     },
     "daily_submit_limit": {
      "anyOf": [
       {
        "type": "integer",
        "maximum": 50.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Daily Submit Limit"
     }
    },
    "type": "object",
    "title": "SourcePatch"
   },
...
  }
 }
}